requirements.txt
pywebview>=5.0.5
pypdfium2>=4.30.0
Pillow>=10.0.0
